fix: 修改Comment-service模块 数据库列名问题, UserID从UserContent获取为Long的问题,path路径传参@错误问题

+++ final-chapters-backend/comment-service/src/main/java/org/ali5669/commentservice/interceptor/UserInfoInterceptor.java
@@ -14,7 +14,7 @@ public class UserInfoInterceptor implements HandlerInterceptor {
         // 2.判断是否为空
         if (StrUtil.isNotBlank(userInfo)) {
             // 不为空，保存到ThreadLocal
-            UserContext.setUserId(Long.valueOf(userInfo));
+            UserContext.setUserId(Integer.valueOf(userInfo));
         }
         // 3.放行
         return true;
